Finally managed to integrate Ed448 to libgcrypt. Test code provided here.

+++ src/utils/utils.cpp
@@ -83,6 +83,6 @@ namespace ecrp {
     }
 
 	uint32_t getTimestampUTC() {
-		return std::time(nullptr);
+		return (uint32_t)std::time(nullptr);
 	}
 }
